feat: reimplement plugin for prettier v2
